requirements.txt
selenium
python-dotenv
sqlalchemy
uvicorn
fastapi
cryptography
pymysql
fastapi[all]